refactoring remote builder
