+++ src/model/index.ts
@@ -10,7 +10,7 @@ import Project from './project';
 import Unity from './unity';
 import Versioning from './versioning';
 import Kubernetes from './kubernetes';
-import AWS from './aws';
+import RemoteBuilder from './remote-builder/remote-builder';
 
 export {
   Action,
@@ -25,5 +25,5 @@ export {
   Unity,
   Versioning,
   Kubernetes,
-  AWS,
+  RemoteBuilder as AWS,
 };
