+++ src/model/remote-builder/remote-builder-alphabet.ts
@@ -1,0 +1,4 @@
+class RemoteBuilderAlphabet {
+  static alphabet = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
+}
+export default RemoteBuilderAlphabet;
