+++ src/model/remote-builder/remote-builder-environment-variable.ts
@@ -1,0 +1,5 @@
+class RemoteBuilderEnvironmentVariable {
+  public name!: string;
+  public value!: string;
+}
+export default RemoteBuilderEnvironmentVariable;
