+++ src/model/remote-builder/remote-builder-secret.ts
@@ -1,0 +1,6 @@
+class RemoteBuilderSecret {
+  public ParameterKey!: string;
+  public EnvironmentVariable!: string;
+  public ParameterValue!: string;
+}
+export default RemoteBuilderSecret;
